+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOp/Meet0Code.java
@@ -197,7 +197,7 @@ import org.firstinspires.ftc.robotcore.external.navigation.CurrentUnit;((DcMotor
 
 } */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.TeleOp;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
@@ -280,4 +280,4 @@ public class Meet0Code extends LinearOpMode {
 
         }
     }
-    }
+}
