+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOp/linearSliderTesting.java
@@ -1,3 +1,5 @@
+package org.firstinspires.ftc.teamcode.TeleOp;
+
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
@@ -20,7 +22,7 @@ public class linearSliderTesting extends LinearOpMode {
         }
     }
 
- 
+
